web: move alerts, markdown and loader input to wildcard
