+++ client/wildcard/src/components/ErrorAlert/index.ts
@@ -1,0 +1,1 @@
+export * from './ErrorAlert'
