+++ client/wildcard/src/components/ErrorMessage/index.ts
@@ -1,0 +1,1 @@
+export * from './ErrorMessage'
