+++ client/wildcard/src/components/Form/LoaderInput/index.ts
@@ -1,0 +1,1 @@
+export * from './LoaderInput'
