+++ client/wildcard/src/components/Markdown/index.ts
@@ -1,0 +1,1 @@
+export * from './Markdown'
