+++ client/wildcard/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { screenReaderAnnounce, screenReaderClearAnnouncements } from './screenReaderAnnounce'
+export { createLinkClickHandler } from './linkClickHandler'
