+++ client/wildcard/src/utils/linkClickHandler.ts
